requirements.txt
# filepath: f:\Programming\Vibecoding\journey_app\requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-APScheduler==1.13.0
python-dotenv==1.0.0
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2